refactor: extract duplicate utility
