+++ apps/server/src/utils/__tests__/parser.test.js
@@ -1,7 +1,7 @@
 import { vi } from 'vitest';
 import { dbModel } from '../../models/dataModel.ts';
 import { parseExcel, parseJson, validateEvent } from '../parser.ts';
-import { makeString, validateDuration } from '../parserUtils.ts';
+import { makeString } from '../parserUtils.ts';
 import { parseAliases, parseUserFields, parseViewSettings } from '../parserFunctions.ts';
 import { EndAction, TimerType } from 'ontime-types';
 import { dayInMs } from 'ontime-utils';
@@ -897,35 +897,5 @@ describe('test views import', () => {
     };
     const parsed = parseViewSettings(testData, true);
     expect(parsed).toStrictEqual(expectedParsedViewSettings);
-  });
-});
-
-describe('test validateDuration()', () => {
-  describe('handles valid inputs', () => {
-    const valid = [
-      { test: 'zero values', timeStart: 0, timeEnd: 0, expected: 0 },
-      { test: 'end after start', timeStart: 0, timeEnd: 1, expected: 1 },
-    ];
-
-    valid.forEach((t) => {
-      it(t.test, () => {
-        const d = validateDuration(t.timeStart, t.timeEnd);
-        expect(d).toBe(t.expected);
-      });
-    });
-  });
-
-  describe('handles edge cases', () => {
-    // edge cases
-    const testData = [
-      { test: 'end the day after', timeStart: 86350000, timeEnd: 3660000, expected: 3660000 + (dayInMs - 86350000) },
-    ];
-
-    testData.forEach((t) => {
-      it(t.test, () => {
-        const d = validateDuration(t.timeStart, t.timeEnd);
-        expect(d).toBe(t.expected);
-      });
-    });
   });
 });
