+++ apps/server/src/utils/__tests__/parserUtils.test.ts
@@ -1,5 +1,4 @@
-import { isEmptyObject, mergeObject, removeUndefined, validateDuration } from '../parserUtils.js';
-import { dayInMs } from 'ontime-utils';
+import { isEmptyObject, mergeObject, removeUndefined } from '../parserUtils.js';
 
 describe('isEmptyObject()', () => {
   test('finds an empty object', () => {
@@ -86,23 +85,5 @@ describe('removeUndefined()', () => {
       third: 'null',
     };
     expect(removeUndefined(obj)).toStrictEqual(obj);
-  });
-});
-
-describe('validateDuration()', () => {
-  it('is the difference between end and start', () => {
-    const duration = validateDuration(10, 20);
-    expect(duration).toBe(10);
-  });
-  it('handles no difference', () => {
-    const duration1 = validateDuration(0, 0);
-    const duration2 = validateDuration(dayInMs, dayInMs);
-    expect(duration1).toBe(0);
-    expect(duration2).toBe(0);
-  });
-  it('handles events that go over midnight', () => {
-    const duration = validateDuration(51, 50);
-    expect(duration).not.toBe(-50);
-    expect(duration).toBe(dayInMs - 1);
   });
 });
